refactor(ai): consolidate Responses replay ownership
